Added Comments from the Java Version
Added several comments from the java version to help understand the
instructions.

In addition have added a slight change to the Board.getPossibleMoves()
just to accommodate the array - however we need to implement a 0 length
array and add to it as we obtain moves. Not sure how to do this - need
to check scala.doc

Once complete address function way of solving this method