Completed MinMax added functional bits
Completed and updated Minimax method using wiki concept - then converted
to a slight pattern matching and functional way of doing things.

the control d != 0 was added to creategametree - as the game would never
call the minimax method.

added the different players to each level of the states.

overhaul of the iterative processes for more functional processes where
possible.